requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
python-telegram-bot==21.7
sniffio==1.3.1
socksio==1.0.0
emoji==2.14.0
requests==2.32.3